pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rocket_ground_station"
version = "25.5.0"
description = "Rocket Ground Station for AGH Space Systems"
readme = "README.md"
requires-python = ">=3.8"
urls = {Homepage = "https://github.com/AGH-Space-Systems-Software/rocket-ground-station.git"}
dependencies = [
    "wheel==0.43.0",
    "kivy==2.3.1",
    "KivyMD==1.1.1",
    "bitstruct==8.19.0",
    "PyYAML==6.0.1",
    "pylint==3.1.0",
    "crc==6.1.1",
    "crccheck==1.3.0",
    "ipython==8",
    "ipython-genutils==0.2.0",
    "prompt-toolkit==3.0.43",
    "pyserial==3.5",
    "colorama==0.4.6",
    "parameterized==0.9.0",
    "kivy-garden==0.1.5",
    "kivy-garden.graph==0.4.0",
    "pandas==2.2.2",
    "docutils==0.20.1; platform_system == 'Windows'",
    "pygments==2.17.2; platform_system == 'Windows'",
    "pypiwin32==223; platform_system == 'Windows'",
    "pygame; platform_system == 'Linux'",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = ["rocket_ground_station*"]

[tool.setuptools.dynamic]
version = {attr = "rocket_ground_station.__version__"}
